refactor(bootstrap): extract multi-tenant builders from dynamic_infrastructure_provider
Split dynamic_infrastructure_provider.go (544 → 422 lines) by extracting
multi-tenant manager construction into dynamic_infrastructure_multi_tenant.go
(138 lines) with dedicated unit tests.

Extracted functions:
- dynamicMultiTenantKey (cache key with hashed API secret)
- multiTenantModeEnabled (config predicate)
- buildCanonicalTenantManager (tenant-manager HTTP client + postgres manager)
- buildRabbitMQTenantManager / buildRabbitMQTenantManagerWithClient

Addresses code review finding M1 (file size > 500 lines).
